Update tests and linter
