+++ CMakeLists.txt
@@ -21,7 +21,7 @@ add_executable(ImageProcessor
         include/change_threads.h
         include/filters/edge_detection.h
         include/filters/gaussian_blur.h
-        include/filters/smoothing_filter.h
+        include/filters/smoothing.h
         src/main.cpp
         src/parameters.cpp
         src/filters/crop.cpp
@@ -35,6 +35,6 @@ add_executable(ImageProcessor
         src/change_threads.cpp
         src/filters/edge_detection.cpp
         src/filters/gaussian_blur.cpp
-        src/filters/smoothing_filter.cpp
+        src/filters/smoothing.cpp
 )
